fix/refactor: correction calculs timebank et gestion des erreurs custom
- Services : Ajout d'un arrondi à 2 décimales pour gommer les imprécisions des flottants sur le solde d'heures (TimeInBank passe en float64).
- Services : Forçage des calculs de dates en UTC strict (getLastMondayUTC) pour éviter les décalages liés au serveur local.
- Repositories : Vérification de  sur  pour retourner un vrai 404 si l'utilisateur n'existe pas (correction d'un faux positif).
- Controllers : Refonte du switch d'erreurs pour exploiter directement les messages précis définis dans .

+++ API/models/DTOs/time_bank_dto.go
@@ -9,6 +9,6 @@ type TimeBankConfigDTO struct {
 
 // Pour afficher le solde (GET root)
 type TimeBankBalanceDTO struct {
-	IsConfigured bool `json:"isConfigured"`
-	TimeInBank   *int `json:"timeInBank"`
+	IsConfigured bool     `json:"isConfigured"`
+	TimeInBank   *float64 `json:"timeInBank"`
 }
